[[ Android ]] Ensure touch event coords are correct
This patch ensures the android touch event coordinates are correct when
there is a disabled scroller control under the touch location. Previously
the ViewGroup containing the scroller manually dispatched the event, but
this caused the coordinates to be relative to the scroller view (and in
particular not taking into account the scrollbar). Instead we return false
so that the event is not consumed, and propagates to the main view with
the correct coordinate parameters.